requirements.txt
python-dotenv==1.2.1
Flask==3.1.0
Flask-SQLAlchemy==3.1.1